pull out threat intel rest tests into separate test class
